reorganize lifecycle components
